tests, libvmi, libpod: Move `GetPodByVirtualMachineInstance` to `libpod`
Lookup a pod based on a VMI is moved from libvmi to libpod where it fits
better.

`libvmi` is focused on spec building and possibly inspection.
