+++ tests/guestlog/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//tests/exec:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
+        "//tests/libpod:go_default_library",
         "//tests/libvmi:go_default_library",
         "//tests/testsuite:go_default_library",
         "//vendor/github.com/google/goexpect:go_default_library",
