+++ tests/libpod/BUILD.bazel
@@ -12,11 +12,13 @@ go_library(
     importpath = "kubevirt.io/kubevirt/tests/libpod",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/controller:go_default_library",
         "//pkg/pointer:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//tests/exec:go_default_library",
         "//tests/flags:go_default_library",
+        "//tests/framework/kubevirt:go_default_library",
         "//tests/testsuite:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
