+++ tests/libvmi/BUILD.bazel
@@ -19,11 +19,9 @@ go_library(
     importpath = "kubevirt.io/kubevirt/tests/libvmi",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/controller:go_default_library",
         "//pkg/pointer:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//tests/containerdisk:go_default_library",
-        "//tests/framework/kubevirt:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
