+++ tests/operator/BUILD.bazel
@@ -28,6 +28,7 @@ go_library(
         "//tests/libmigration:go_default_library",
         "//tests/libnet:go_default_library",
         "//tests/libnode:go_default_library",
+        "//tests/libpod:go_default_library",
         "//tests/libstorage:go_default_library",
         "//tests/libvmi:go_default_library",
         "//tests/libwait:go_default_library",
