+++ tests/vmi_lifecycle_test.go
@@ -384,7 +384,7 @@ var _ = Describe("[rfe_id:273][crit:high][arm64][vendor:cnv-qe@redhat.com][level
 					vmi = tests.RunVMIAndExpectScheduling(vmi, 30)
 					ctx, cancel := context.WithCancel(context.Background())
 					defer cancel()
-					launcher, err := libvmi.GetPodByVirtualMachineInstance(vmi, vmi.Namespace)
+					launcher, err := libpod.GetPodByVirtualMachineInstance(vmi, vmi.Namespace)
 					Expect(err).ToNot(HaveOccurred())
 					watcher.New(launcher).
 						SinceWatchedObjectResourceVersion().
@@ -413,7 +413,7 @@ var _ = Describe("[rfe_id:273][crit:high][arm64][vendor:cnv-qe@redhat.com][level
 					}
 					By("Starting a VirtualMachineInstance")
 					createdVMI := tests.RunVMIAndExpectScheduling(vmi, 30)
-					launcher, err := libvmi.GetPodByVirtualMachineInstance(createdVMI, createdVMI.Namespace)
+					launcher, err := libpod.GetPodByVirtualMachineInstance(createdVMI, createdVMI.Namespace)
 					Expect(err).ToNot(HaveOccurred())
 					// Wait until we see that starting the VirtualMachineInstance is failing
 					By(fmt.Sprintf("Checking that VirtualMachineInstance start failed: starting at %v", time.Now()))
